Extract header files for range and appointment.
